+++ docs/changelog.rst
@@ -17,6 +17,11 @@ Next
   only data greater than ``thresh`` is plotted.
 * Mitigate pending deprecated imports from ``collections`` - :issue:`313`
 
+Bug fixes
+---------
+
+* Tally data for detectors with time-bins are properly handled - :issue:`312`
+
 .. _v0.7.0:
 
 :release-tag:`0.7.0`